+++ src/ndarray.jl
@@ -99,7 +99,7 @@ type NDArray
 end
 
 function Base.show(io :: IO, arr :: NDArray)
-  println(io, "$(join(size(arr), "x")) mx.NDArray{$(eltype(arr))} @ $(context(arr)):")
+  println(io, "$(join(size(arr), "×")) mx.NDArray{$(eltype(arr))} @ $(context(arr)):")
   Base.showarray(io, try_get_shared(arr, sync=:read), false, header=false)
 end
 
@@ -971,30 +971,105 @@ function save(filename::String, data::Dict{Base.Symbol,NDArray})
           filename, length(names), arrays, names)
 end
 
-import Base: reshape
+################################################################################
+# Mapping NDArray functions to Base-like API
+################################################################################
 
-"""
-    reshape(arr::NDArray, dim...; reverse=false)
-    reshape(arr::NDArray, dim; reverse=false)
-"""
-reshape{N}(arr::NDArray, dim::NTuple{N, Integer}; reverse::Bool=false) =
-    _reshape(arr, dim, reverse)
-reshape{N}(arr::NDArray, dim::Vararg{Integer, N}; reverse::Bool=false) =
-    _reshape(arr, dim, reverse)
+const _mxsig = Dict{Symbol,Expr}()
 
-@inline function _reshape{N}(arr::NDArray, dim::NTuple{N, Integer}, reverse::Bool)
-  op_handle = _get_cached_libmx_op_handle("reshape")
-  n_output = Ref(Cint(0))
-  hdls_ref = Ref{Ptr{MX_handle}}(C_NULL)
-  @mxcall(:MXImperativeInvoke,
-          (MX_handle, Cint, Ptr{MX_handle}, Ref{Cint}, Ref{Ptr{MX_handle}},
-           Cint, char_pp, char_pp),
-          op_handle, 1, [arr.handle], n_output, hdls_ref,
-          2, ["shape", "reverse"], [dump_mx_param(dim), dump_mx_param(!reverse)])
-  #                                                       not a typo  ^^^^^^^^
-  @assert n_output[] == 1
-  NDArray(MX_NDArrayHandle(unsafe_load(hdls_ref[], 1)))
+function _autoimport(name::Symbol)
+  if isdefined(Base, name)
+    :(import Base: $name)
+  else
+    :()
+  end
 end
+
+macro _remap(sig::Expr, imp::Expr)
+  fname = sig.args[1]
+  opname = string(imp.args[1])
+
+  import_expr = _autoimport(fname)
+
+  if isa(imp.args[2], Expr) && imp.args[2].head == :parameters
+    ndin = imp.args[3:end]
+    mxargs = imp.args[2].args
+  else  # no keyword arguments
+    ndin = imp.args[2:end]
+    mxargs = []
+  end
+
+  mxkeys = map(x -> string(x.args[1]), mxargs)
+  mxvals = Expr(:vect, map(x -> :(dump_mx_param($(x.args[2]))), mxargs)...)
+  ndhlds = Expr(:vect, map(x -> :($(x).handle), ndin)...)
+
+  func_body = quote
+    op_handle = _get_cached_libmx_op_handle($opname)
+    n_output = Ref(Cint(0))
+    hdls_ref = Ref{Ptr{MX_handle}}(C_NULL)
+    @mxcall(:MXImperativeInvoke,
+            (MX_handle,
+             Cint,
+             Ptr{MX_handle},
+             Ref{Cint},
+             Ref{Ptr{MX_handle}},
+             Cint,
+             char_pp,
+             char_pp),
+            op_handle,
+            $(length(ndin)),
+            $(ndhlds),
+            n_output,
+            hdls_ref,
+            $(length(mxargs)),
+            $mxkeys,
+            $mxvals)
+    NDArray(MX_NDArrayHandle(unsafe_load(hdls_ref[], 1)))
+  end
+
+  docstr = "    $sig"
+  func_def = Expr(:function, sig, func_body)
+
+  esc(quote
+    $import_expr
+    @doc $docstr ->
+    $func_def
+  end)
+end
+
+macro _remap(sig::Expr, imp::Symbol)
+  imp = _mxsig[imp]
+
+  esc(quote
+    @_remap($sig, $imp)
+  end)
+end
+
+_mxsig[:reshape] = :(reshape(arr; shape = dim, reverse = !reverse))
+@_remap reshape(arr::NDArray, dim...; reverse = false) reshape
+@_remap reshape(arr::NDArray, dim; reverse = false)    reshape
+
+@_remap mean(arr::NDArray)         mean(arr)
+@_remap mean(arr::NDArray, region) mean(arr; axis = 0 .- region, keepdims = true)
+
+@_remap sum(arr::NDArray)       sum(arr)
+@_remap sum(arr::NDArray, dims) sum(arr; axis = 0 .- dims, keepdims = true)
+
+@_remap maximum(arr::NDArray)       max(arr)
+@_remap maximum(arr::NDArray, dims) max(arr; axis = 0 .- dims, keepdims = true)
+
+@_remap minimum(arr::NDArray)       min(arr)
+@_remap minimum(arr::NDArray, dims) min(arr; axis = 0 .- dims, keepdims = true)
+
+# See https://github.com/dmlc/MXNet.jl/issues/55
+@_remap dot(x::NDArray, y::NDArray) dot(y, x)
+
+# See https://github.com/dmlc/MXNet.jl/pull/123
+@_remap transpose(arr::NDArray) transpose(_only2d(arr))
+@_remap permutedims(arr::NDArray, axes) transpose(arr; axes = length(axes) .- tuple(axes...))
+
+@_remap prod(arr::NDArray)       prod(arr)
+@_remap prod(arr::NDArray, dims) prod(arr; axis = 0 .- dims, keepdims = true)
 
 ################################################################################
 # NDArray functions dynamically imported from libmxnet
@@ -1063,19 +1138,6 @@ function _get_ndarray_function_def(name :: String)
         args = MX_handle[]
       end
 
-      # XXX: hacky way of solving the problem that the arguments of `dot` should be swapped
-      # See https://github.com/dmlc/MXNet.jl/issues/55
-      if $name == "dot"
-        args = reverse(args)
-      end
-
-      # XXX: hacky way of solving the semantic difference of the axes parameter in Julia
-      # and in libmxnet.
-      # See https://github.com/dmlc/MXNet.jl/pull/123
-      if $name == "transpose"
-        kwargs = Any[key != :axes ? (key, arg) : (key, map(i->length(arg)-i, arg)) for (key, arg) in kwargs]
-      end
-
       if length(output_vars) > 0
         output_handles = map((x) -> Base.cconvert(MX_handle, x), output_vars)
         # XXX: Julia 0.4 has bug: [Array{MX_handle}] == Array{MX_handle}
@@ -1123,9 +1185,21 @@ function _get_ndarray_function_def(name :: String)
   return func_def, func_def2
 end
 
+const _op_import_bl = [  # import black list; do not import these funcs
+    "mean",
+    "reshape",
+    "sum",
+    "max",
+    "max_axis",
+    "min",
+    "min_axis",
+    "dot",
+    "transpose",
+    "prod",
+]
+
 macro _import_ndarray_functions()
-  black_list = ["reshape"]  # do not import these funcs
-  names = filter(n -> ∉(lowercase(n), black_list), _get_libmx_op_names())
+  names = filter(n -> ∉(lowercase(n), _op_import_bl), _get_libmx_op_names())
 
   func_exprs = map(names) do name
     op_handle = _get_libmx_op_handle(name)